requirements.txt
flask
flask_sqlalchemy
pymysql
flask_migrate
cryptography